requirements.txt
click==8.1.7
dash==2.14.2
datasets==2.16.1
fastapi==0.109.0
happytransformer==3.0.0
hydra-core==1.3.2
Markdown==3.5.1
matplotlib==3.8.2
numpy==1.26.3
omegaconf==2.3.0
prometheus_fastapi_instrumentator==6.1.0
pydantic==2.5.3
pytest==7.4.4
pytorch_lightning==2.1.3
Requests==2.31.0
scikit_learn==1.3.2
torch==2.1.2
